+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.1.0] - Not yet released
+### New
+- Multi-Domain support (aka multi-tenant), each domain can server a specific host url with independent configuration and management 
+### Changed
+- Removed setting 'Security ⇨ Web Security ⇨ Permitted IP Network Addresses', this functionality is better provided by the web server itself.
+
 ## [2.0.1] - Released March 11, 2022
 ### Changed
 - Issue #573 - PWM 5081 at the end of user activation ( no profile assigned )